% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHEX_with_generateImage4R.R
\name{createHEX_with_generateImage4R}
\alias{createHEX_with_generateImage4R}
\title{Create Hex Stickers for R Packages}
\usage{
createHEX_with_generateImage4R(description, package_name, n = 1)
}
\arguments{
\item{description}{A string. This is the description of the R package for which the stickers will be generated.}

\item{package_name}{A string. This is the name of the R package for which the stickers will be generated.}

\item{n}{An integer. This is the number of stickers to generate. The default value is 3.}
}
\value{
A list of EBImage objects. These are the hexadecimal stickers generated for the R package.
}
\description{
Generates hexadecimal stickers for R packages using a provided
   description and package name. The stickers are generated by the OpenAI API.
   This is an experimental function.
}
\details{
Create Hex Stickers for R Packages
}
\examples{
\dontrun{
# Define a description for the R package
description = "chatAI4R provides a seamless interface for integrating OpenAI and other APIs into R."
package_name = "chatAI4R"

# Generate hex stickers
res <- createHEX_with_generateImage4R(description, package_name)

# Display the stickers
Display(res)
}
}
\author{
Satoshi Kume
}
