% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDalleImage4R.R
\name{generateDalleImage4R}
\alias{generateDalleImage4R}
\title{Generate Text to Images Using OpenAI DALLE2 API}
\usage{
generateDalleImage4R(
  content,
  model = "dall-e-3",
  n = 1,
  size = "1024x1024",
  response_format = "url",
  Output_image = TRUE,
  SaveImg = FALSE,
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{content}{A character string, the content to generate images.}

\item{model}{A model name, dall-e-3 or dall-e-2.}

\item{n}{A count, the number of images to generate. For dall-e-2, Must be between 1 and 10. For dall-e-3, only n=1 is supported.}

\item{size}{A character string, the size of the images. Must be one of "256x256", "512x512", "1024x1024".}

\item{response_format}{A character string, the format of the response from the API. Must be one of "url", "b64_json".}

\item{Output_image}{A flag, whether to output the images.}

\item{SaveImg}{A flag, whether to save the images.}

\item{api_key}{A character string, the API key for OpenAI.}
}
\value{
If Output_image is TRUE, returns a list of images; if Output_image is FALSE, returns a character vector of URLs.
}
\description{
This function generates images from text input using OpenAI DALLE 2/3 API.
}
\details{
Generate Text to Images Using DALLE 2/3 API
}
\examples{
\dontrun{
# Generate images
content = "Japanese something"
results <- generateDalleImage4R(content, n = 1)
# Display the generated images
Display(results)
}
}
\author{
Satoshi Kume
}
