% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2img4HEX.R
\docType{data}
\name{text_prompts}
\alias{text_prompts}
\title{Create Hex Stickers using text-to-image via Stable Diffusion R Packages}
\format{
An object of class \code{character} of length 1.
}
\usage{
text_prompts
}
\description{
Create Hex Stickers using text-to-image via Stable Diffusion R Packages
}
\keyword{datasets}
