% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2JPanime_v01.R
\name{txt2JPanime_v01}
\alias{txt2JPanime_v01}
\title{Text-to-image Japanese anime generator v01 using Stable Diffusion}
\usage{
txt2JPanime_v01(
  text_prompts = "",
  weight = 0.5,
  height = 512,
  width = 512,
  number_of_images = 1,
  steps = 10,
  cfg_scale = 7,
  clip_guidance_preset = "NONE",
  sampler = "NONE",
  seed = 0,
  style_preset = "NONE",
  engine_id = "stable-diffusion-512-v2-1",
  api_host = "https://api.stability.ai",
  api_key = Sys.getenv("DreamStudio_API_KEY"),
  verbose = TRUE
)
}
\description{
Text-to-image Japanese anime generator v01 using Stable Diffusion
}
